snmp: extract ASN.1 BER codec from agent.go
Pure code-motion split of the over-threshold pkg/snmp/agent.go (2143
prod LOC) as partial progress on the #5661 Go control-plane modularity
cohort. The SNMP wire codec is a self-contained layer of Agent-
independent package-level functions, cleanly separable from the agent
state machine, packet handlers, and MIB view.

New pkg/snmp/agent_ber.go holds, byte-identical:
  - oidHasPrefix and the OID comparison helpers (oidEqual, oidCompare)
  - the value encoders (berEncodeCounter32/Gauge32/Counter64,
    berEncodeTimeTicks, berEncodeOID/SubID/Value, tagTimeTicks)
  - the TLV/length/integer encoders (berEncodeTLV, berEncodeLength,
    berEncodeIntegerTLV, berEncodeIntegerValue)
  - the decoders (berDecodeHeader/Length/Integer/OctetString/OID,
    berEncodedLen) and decodePDUFields

None of these have an *Agent receiver, so the move is a clean boundary
with no coupling change. No declarations renamed, no logic changed, same
package so no import cycle. agent.go drops from 2143 to 1737 prod LOC,
under the ~2000 modularity threshold; its now-unused "encoding/binary"
import is removed. go build ./... and go test ./pkg/snmp/... pass;
gofmt clean.

Partial progress on #5661 (low-risk routing/snmp/cli files); HA-critical
files (daemon/vrrp/cluster) tracked separately.
